+++ docs/CHANGELOG.md
@@ -1,1 +1,3 @@
 ## Changelog for v2.2.3
+
+- Forgot to declare FGB driver(W/R) as function